Working?

+++ editJournal.php
@@ -17,11 +17,12 @@ if ($result->num_rows > 0) {
 $site_url = site_url();
 
 ?>
-<main>
 <a href="journal.php" class="label_back">
         <img class="label_back_arrow" src="media/icons/back.svg">
         <p class=" BS label_back_text">Journal</p>
         </a>
+<main>
+
 <div class="main_label">
             <div class="main_label_header">
                 <img class="icon main_label_icon" src="media/icons/edit.svg"/>
@@ -41,7 +42,7 @@ $site_url = site_url();
       <label for="title">Text</label> 
       <textarea class="journal_text" type="text" id="entryText" name="entryText"><?php echo $entry['entryText']?></textarea>
 </div>
-    <input class="submit_button" type="submit" value="submit" class="editSubmit">
+    <input type="submit" value="Submit" class="editSubmit TS">
     <input type="hidden" name="EntryDate" value="<?php echo $entry['EntryDate']?>">
     <input type="hidden" name="id" value="<?php echo $entry['id']?>">
     </form>
@@ -50,13 +51,13 @@ $site_url = site_url();
     </div>
 
     <div class="delete_confirmation js-confirm hidden">
-      <h3 class="TS">Are you sure you want to delete your entry: <?php echo $entry['title']?> ?</h3>
-      <a href="<?php echo $site_url?>/includes/deleteJournal.php?id=<?php echo $entry['id']?>">
-        <h3 class="TS">Yes</h3>
-      </a>
-      <div class="confirmation_button_no js-cancel">
+      <h3 class="TS confirmation_text">Are you sure you want to delete this journal entry?</h3>
+      <div class="confirmation_button confirmation_button_no js-cancel">
         <h3 class="TS">No</h3>
       </div>
+      <a class="confirmation_button confirmation_button_yes" href="<?php echo $site_url?>/includes/deleteJournal.php?id=<?php echo $entry['id']?>">
+        <h3 class="TS">Yes</h3>
+      </a>
     </div>
   </main>
   <script src="scripts/deleteJournal.js"></script>
